requirements.txt
fastapi>=0.104.1
uvicorn==0.24.0
json5>=0.9.0
python-multipart==0.0.6
pydantic==2.5.2
openai>=1.26.0,<2.0.0
python-dotenv==1.0.0
jinja2==3.1.2
google-genai==0.6.0
packaging>=23.2,<24.0 
langchain-core==0.3
langchain-openai==0.2.1
google-cloud-aiplatform>=1.36.0 
google-generativeai>=0.3.0
 